Подготовка отчета к базовому выводу МТО(на случай бэкапа))

+++ ElectroMod/Dtos/TransformatorContainerDto.cs
@@ -11,6 +11,6 @@ namespace ElectroMod.DataBase.Dtos
         public string TypeKTP { get; set; }
         public double Uk { get; set; }
         public double Pk { get; set; }
-        public double S { get; set; }
+        public int S { get; set; }
     }
 }
